move transaction types to sperated file
